Refactor cesium store
